Upgrade the template code snippets to v3.0.0
The tutorial in "Prepare to build a dApp" section mentions "You should already have version v3.0.0 of the Substrate Developer Hub Node Template", but seems the code snippets are still from v2.0.0 which are not compatible with v3.0.0 and makes the pallet template fail to compile for node template v3.0.0. Apparently the frame_system::Trait module has been renamed to  frame_system::Config in v3.0.0.