Update tutorial to match Nicks balances weight type implimentation (#997)
* Updated to match code in repo

Weight information for is specified as pallet_balances::weights::SubstrateWeight<Runtime> in code for Balances at runtime/src/lib.rs (~L234), not () as stated in this current file. This appears to be a fairly recent change, and is not reflected in either v1 or v2 of the Nicks pallet as shown currently here: https://substrate.dev/docs/en/knowledgebase/runtime/pallets in contrast. This might be confusing. The wording or comparison may also be affected.

* Update docs/tutorials/add-a-pallet/configure-a-pallet.md

Co-authored-by: Dan Shields <35669742+NukeManDan@users.noreply.github.com>